08-task/example/gigachat_api/requirements.txt
gigachat
python-dotenv
streamlit

